Rename decision_scene_option->decision_scene
